feat: add handle for '#no_sync' tag handle

+++ mtSync.py
@@ -304,6 +304,12 @@ def sync_main(toot_id):
         skip_toot = False # 重置跳过嘟文标记
         return 0
 
+    # 检查嘟文结尾是否以 “#no_sync” tag 结尾
+    if toot_text.endswith('#no_sync'):
+        tprint(colored('[Check] 嘟文结尾包含不同步标签 #no_sync ，跳过...','green'))
+        save_synced_toots(toot_id)
+        return 0
+
     # 清空媒体缓存文件夹
     if os.path.exists('./media/'):
         shutil.rmtree('./media/',ignore_errors=True)
